@@ -1,39 +1,47 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App;
 
 use App\FrontMatter\ParserInterface;
 use RuntimeException;
+
+use function file_exists;
+use function file_put_contents;
+use function glob;
+use function uasort;
+use function var_export;
+
+use const LOCK_EX;
 
 abstract class AbstractCollection
 {
     protected const FOLDER_COLLECTION = '';
     protected const CACHE_FILE        = '';
 
-    protected $collection = [];
+    protected array $collection = [];
 
-    /** @var ParserInterface */
-    protected $frontMatterParser;
-
-    public function __construct(ParserInterface $frontMatterParser)
+    public function __construct(protected ParserInterface $frontMatterParser)
     {
         if (empty(static::CACHE_FILE)) {
             throw new RuntimeException('The cache file path is not defined!');
         }
-        $this->frontMatterParser = $frontMatterParser;
+
         if (! file_exists(static::CACHE_FILE)) {
             $this->buildCache();
-        } else {
-            $this->collection = require static::CACHE_FILE;
+            return;
         }
+
+        $this->collection = require static::CACHE_FILE;
     }
 
-    public function getAll()
+    public function getAll(): array
     {
         return $this->collection;
     }
 
-    public function getFromFile($file)
+    public function getFromFile(string $file): array
     {
         $result = [];
         if (file_exists($file)) {
@@ -44,22 +52,23 @@ abstract class AbstractCollection
         return $result;
     }
 
-    protected function buildCache()
+    protected function buildCache(): void
     {
         if (empty(static::FOLDER_COLLECTION)) {
             throw new RuntimeException('The folder collection is not defined!');
         }
 
         foreach (glob(static::FOLDER_COLLECTION . '/*.md') as $file) {
-            $doc = $this->frontMatterParser->parse($file);
-            $fields = $doc->getYAML();
+            $doc                     = $this->frontMatterParser->parse($file);
+            $fields                  = $doc->getYAML();
             $this->collection[$file] = $fields;
         }
+
         uasort($this->collection, [$this, 'order']);
         file_put_contents(static::CACHE_FILE, '<?php return ' . var_export($this->collection, true) . ';', LOCK_EX);
     }
 
-    protected function order($a, $b)
+    protected function order(array $a, array $b): bool|int
     {
         return false;
     }
