@@ -1,4 +1,5 @@
-<?php
+<?php // phpcs:disable Generic.Files.LineLength.TooLong
+
 
 declare(strict_types=1);
 
