@@ -1,10 +1,11 @@
-<?php
+<?php // phpcs:disable PSR12.Files.FileHeader.IncorrectOrder
+
 
 declare(strict_types=1);
 
-use Psr\Container\ContainerInterface;
 use Mezzio\Application;
 use Mezzio\MiddlewareFactory;
+use Psr\Container\ContainerInterface;
 
 /**
  * Setup routes with a single request method:
@@ -32,8 +33,8 @@ use Mezzio\MiddlewareFactory;
  *     'contact'
  * );
  */
-return function (Application $app, MiddlewareFactory $factory, ContainerInterface $container) : void {
 
+return function (Application $app, MiddlewareFactory $factory, ContainerInterface $container): void {
     // Application routes
     (new App\ConfigProvider())->registerRoutes($app, '/');
 
