@@ -1,19 +1,17 @@
 <?php
 
-/**
- * @license http://opensource.org/licenses/BSD-2-Clause BSD-2-Clause
- * @copyright Copyright (c) Matthew Weier O'Phinney
- */
-
 declare(strict_types=1);
 
 namespace GetLaminas\Blog\Console;
 
 use GetLaminas\Blog\Mapper\MapperInterface;
-use Psr\Container\ContainerInterface;
 use Mezzio\Helper\ServerUrlHelper;
 use Mezzio\Router\RouterInterface;
 use Mezzio\Template\TemplateRendererInterface;
+use Psr\Container\ContainerInterface;
+
+use function getcwd;
+use function realpath;
 
 class FeedGeneratorFactory
 {
