@@ -1,10 +1,5 @@
 <?php
 
-/**
- * @license http://opensource.org/licenses/BSD-2-Clause BSD-2-Clause
- * @copyright Copyright (c) Matthew Weier O'Phinney
- */
-
 declare(strict_types=1);
 
 namespace GetLaminas\Blog\Console;
@@ -19,6 +14,7 @@ use SplFileInfo;
 
 use function is_dir;
 use function is_string;
+use function sprintf;
 
 /**
  * Usage:
