@@ -1,14 +1,11 @@
 <?php
 
-/**
- * @license http://opensource.org/licenses/BSD-2-Clause BSD-2-Clause
- * @copyright Copyright (c) Matthew Weier O'Phinney
- */
+declare(strict_types=1);
 
 namespace GetLaminas\Blog\Handler;
 
-use Psr\Container\ContainerInterface;
 use Mezzio\Handler\NotFoundHandler;
+use Psr\Container\ContainerInterface;
 
 class FeedHandlerFactory
 {
