@@ -1,16 +1,13 @@
 <?php
 
-/**
- * @license http://opensource.org/licenses/BSD-2-Clause BSD-2-Clause
- * @copyright Copyright (c) Matthew Weier O'Phinney
- */
+declare(strict_types=1);
 
 namespace GetLaminas\Blog\Handler;
 
 use GetLaminas\Blog\Mapper\MapperInterface;
-use Psr\Container\ContainerInterface;
 use Mezzio\Router\RouterInterface;
 use Mezzio\Template\TemplateRendererInterface;
+use Psr\Container\ContainerInterface;
 
 class ListPostsHandlerFactory
 {
