@@ -1,10 +1,5 @@
 <?php
 
-/**
- * @license http://opensource.org/licenses/BSD-2-Clause BSD-2-Clause
- * @copyright Copyright (c) Matthew Weier O'Phinney
- */
-
 declare(strict_types=1);
 
 namespace GetLaminas\Blog\Listener;
@@ -15,9 +10,7 @@ use GetLaminas\Blog\Mapper\MapperInterface;
 
 class FetchBlogPostFromMapperListener
 {
-    /**
-     * @var MapperInterface
-     */
+    /** @var MapperInterface */
     private $mapper;
 
     public function __construct(MapperInterface $mapper)
