@@ -8,37 +8,30 @@ use DateTimeImmutable;
 use Exception;
 use Laminas\Feed\Reader\Reader;
 use Laminas\Feed\Writer\Feed;
-use League\CommonMark\CommonMarkConverter;
+use League\CommonMark\MarkdownConverterInterface;
+use Mezzio\ProblemDetails\ProblemDetailsResponseFactory;
+use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Mezzio\ProblemDetails\ProblemDetailsResponseFactory;
-use Psr\Http\Message\ResponseFactoryInterface;
+
+use function explode;
+use function file_get_contents;
+use function file_put_contents;
+use function parse_url;
+use function sprintf;
+
+use const LOCK_EX;
+use const PHP_URL_PATH;
 
 class ReceiveFeedItemHandler implements RequestHandlerInterface
 {
-    /** @var string */
-    private $feedFile;
-
-    /** @var CommonMarkConverter */
-    private $markdown;
-
-    /** @var ProblemDetailsResponseFactory */
-    private $problemFactory;
-
-    /** @var ResponseFactoryInterface */
-    private $responseFactory;
-
     public function __construct(
-        string $feedFile,
-        CommonMarkConverter $markdown,
-        ResponseFactoryInterface $responseFactory,
-        ProblemDetailsResponseFactory $problemFactory
+        private string $feedFile,
+        private MarkdownConverterInterface $markdown,
+        private ResponseFactoryInterface $responseFactory,
+        private ProblemDetailsResponseFactory $problemFactory
     ) {
-        $this->feedFile        = $feedFile;
-        $this->markdown        = $markdown;
-        $this->responseFactory = $responseFactory;
-        $this->problemFactory  = $problemFactory;
     }
 
     public function handle(ServerRequestInterface $request): ResponseInterface
@@ -93,7 +86,7 @@ class ReceiveFeedItemHandler implements RequestHandlerInterface
             $data['package'],
             $data['version'],
             $data['url'],
-            $this->markdown->convertToHtml($data['changelog']),
+            $this->markdown->convertToHtml($data['changelog'])->getContent(),
             new DateTimeImmutable($data['publication_date']),
             new Author($data['author_name'], $authorUrl)
         );
@@ -106,8 +99,8 @@ class ReceiveFeedItemHandler implements RequestHandlerInterface
         $releases = new Releases();
 
         foreach ($feed as $entry) {
-            $title = $entry->getTitle();
-            list($package, $version) = explode(' ', $title);
+            $title               = $entry->getTitle();
+            [$package, $version] = explode(' ', $title);
 
             $author     = $entry->getAuthor();
             $authorName = $author['name'];
