@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GetLaminas\Security;
 
 use App\AbstractCollection;
@@ -9,7 +11,7 @@ class Advisory extends AbstractCollection
     protected const FOLDER_COLLECTION = 'data/advisories';
     protected const CACHE_FILE        = 'var/advisories.php';
 
-    protected function order($a, $b)
+    protected function order(array $a, array $b): int
     {
         if ($b['date'] === $a['date']) {
             return $b['title'] <=> $a['title'];
