@@ -1,16 +1,26 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GetLaminas\Security\Handler;
 
 use DateTimeImmutable;
 use GetLaminas\Security\Advisory;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Server\RequestHandlerInterface;
-use Psr\Http\Message\ServerRequestInterface;
 use Laminas\Diactoros\Response\HtmlResponse;
 use Laminas\Diactoros\Response\TextResponse;
-use Mezzio\Template;
 use Laminas\Feed\Writer\Feed;
+use Mezzio\Template;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+
+use function array_slice;
+use function basename;
+use function ceil;
+use function count;
+use function current;
+use function preg_match;
+use function sprintf;
 
 class SecurityHandler implements RequestHandlerInterface
 {
@@ -40,20 +50,20 @@ class SecurityHandler implements RequestHandlerInterface
         $content = [];
         if ('advisories' === $action) {
             $params = $request->getQueryParams();
-            $page = isset($params['page']) ? (int) $params['page'] : 1;
+            $page   = isset($params['page']) ? (int) $params['page'] : 1;
 
             $allAdvisories = $this->advisory->getAll();
-            $totPages = (int) ceil(count($allAdvisories) / self::ADVISORY_PER_PAGE);
+            $totPages      = (int) ceil(count($allAdvisories) / self::ADVISORY_PER_PAGE);
 
             if ((0 !== $totPages) && ($page > $totPages || $page < 1)) {
                 return new HtmlResponse($this->template->render('error::404'));
             }
 
-            $nextPage = ($page === $totPages) ? 0 : $page + 1;
-            $prevPage = ($page === 1) ? 0 : $page - 1;
+            $nextPage = $page === $totPages ? 0 : $page + 1;
+            $prevPage = $page === 1 ? 0 : $page - 1;
 
             $advisories = array_slice($allAdvisories, ($page - 1) * self::ADVISORY_PER_PAGE, self::ADVISORY_PER_PAGE);
-            $content = [
+            $content    = [
                 'advisories' => $advisories,
                 'tot'        => $totPages,
                 'page'       => $page,
@@ -64,10 +74,10 @@ class SecurityHandler implements RequestHandlerInterface
         return new HtmlResponse($this->template->render(sprintf('security::%s', $action), $content));
     }
 
-    protected function feed(ServerRequestInterface $request)
+    protected function feed(ServerRequestInterface $request): ResponseInterface
     {
-        $baseUrl = (string) $request->getUri()->withPath('/security');
-        $feedUrl = (string) $request->getUri()->withQuery('')->withFragment('');
+        $baseUrl     = (string) $request->getUri()->withPath('/security');
+        $feedUrl     = (string) $request->getUri()->withQuery('')->withFragment('');
         $advisoryUrl = $request->getUri()->withPath('/security/advisory');
 
         $matches = [];
@@ -93,7 +103,7 @@ class SecurityHandler implements RequestHandlerInterface
 
         foreach ($advisories as $id => $advisory) {
             $content = $this->advisory->getFromFile($id);
-            $entry = $feed->createEntry();
+            $entry   = $feed->createEntry();
             $entry->setTitle($content['title']);
             $entry->setLink(sprintf('%s/%s', $advisoryUrl, basename($id, '.md')));
             $entry->addAuthor([
